Move tree creation to a separate file
