+++ earth_enterprise/src/fusion/autoingest/StorageManager.h
@@ -24,6 +24,7 @@
 #include <memory>
 
 #include "common/khCache.h"
+#include "common/khFileUtils.h"
 #include "common/notify.h"
 #include "common/SharedString.h"
 #include "StorageManagerAssetHandle.h"
