+++ earth_enterprise/src/fusion/autoingest/sysman/SConscript
@@ -97,7 +97,7 @@ gesystemmanager = env.executable('gesystemmanager',
     map(lambda p: 'plugins/'+p+'AssetD.cpp', plugins) +
     [ 'gesystemmanager_main.cpp',
       'AssetD.cpp', 'AssetVersionD.cpp',
-      'AssetOperation.cpp', 'StateUpdater.cpp',
+      'AssetOperation.cpp', 'StateUpdater.cpp', 'DependentStateTree.cpp',
       'khSystemManager.cpp',
       'khAssetManager.cpp', 'khAssetManagerDispatch.cpp',
       'khResourceManager.cpp', 'khTask.cpp', 'TaskRequirements.cpp',
@@ -145,7 +145,7 @@ env.test('LeafAssetVersion_unittest',
     map(lambda p: 'plugins/'+p+'AssetD.o', plugins) +
     [ 'LeafAssetVersion_unittest.cpp',
       'AssetD.o', 'AssetVersionD.o',
-      'AssetOperation.o', 'StateUpdater.o',
+      'AssetOperation.o', 'StateUpdater.o', 'DependentStateTree.o',
       'khSystemManager.o',
       'khAssetManager.o', 'khAssetManagerDispatch.o',
       'khResourceManager.o', 'khTask.o', 'TaskRequirements.o',
@@ -164,7 +164,7 @@ env.test('StateUpdater_unittest',
     map(lambda p: 'plugins/'+p+'AssetD.o', plugins) +
     [ 'StateUpdater_unittest.cpp',
       'AssetD.o', 'AssetVersionD.o',
-      'AssetOperation.o', 'StateUpdater.o',
+      'AssetOperation.o', 'StateUpdater.o', 'DependentStateTree.o',
       'khSystemManager.o',
       'khAssetManager.o', 'khAssetManagerDispatch.o',
       'khResourceManager.o', 'khTask.o', 'TaskRequirements.o',
