+++ earth_enterprise/src/fusion/autoingest/sysman/StateUpdater.h
@@ -14,17 +14,14 @@
  * limitations under the License.
  */
 
-#ifndef ASSETTREE_H
-#define ASSETTREE_H
+#ifndef STATEUPDATER_H
+#define STATEUPDATER_H
 
-#include <boost/graph/adjacency_list.hpp>
-#include <functional>
 #include <set>
 #include <map>
 
+#include "DependentStateTree.h"
 #include "AssetVersion.h"
-#include "autoingest/.idl/storage/AssetDefs.h"
-#include "common/SharedString.h"
 #include "StorageManager.h"
 
 // This class efficiently updates the states of lots of asset versions at
@@ -37,60 +34,13 @@
 class StateUpdater
 {
   private:
-    struct AssetVertex {
-      SharedString name;
-      AssetDefs::State state;
-      bool inDepTree;
-      bool recalcState;
-      bool stateChanged;
-      size_t index; // Used by the dfs function
-    };
-
-    enum DependencyType { INPUT, CHILD, DEPENDENT, DEPENDENT_AND_CHILD };
-
-    struct AssetEdge {
-      DependencyType type;
-    };
-
-    typedef boost::adjacency_list<
-        boost::setS,
-        boost::setS,
-        boost::directedS,
-        AssetVertex,
-        AssetEdge> TreeType;
-    typedef std::map<SharedString, TreeType::vertex_descriptor> VertexMap;
-
-    // Helper struct for building the tree
-    struct TreeBuildData;
-
-    // Used by dfs function to update states of assets in the tree
-    friend struct boost::property_map<TreeType, boost::vertex_index_t>;
     class SetStateVisitor;
 
     StorageManagerInterface<AssetVersionImpl> * const storageManager;
-    TreeType tree;
+    DependentStateTree tree;
 
-    inline bool IsDependent(DependencyType type) { return type == DEPENDENT || type == DEPENDENT_AND_CHILD; }
-
-    void BuildDependentTree(
-        const SharedString & ref,
-        std::function<bool(AssetDefs::State)> includePredicate);
-    TreeType::vertex_descriptor AddOrUpdateVertex(
-        const SharedString & ref,
-        TreeBuildData & buildData,
-        bool inDepTree,
-        bool recalcState,
-        std::set<TreeType::vertex_descriptor> & toFillIn);
-    void FillInVertex(
-        TreeType::vertex_descriptor myVertex,
-        TreeBuildData & buildData,
-        std::set<TreeType::vertex_descriptor> & toFillIn);
-    void AddEdge(
-        TreeType::vertex_descriptor from,
-        TreeType::vertex_descriptor to,
-        AssetEdge data);
     void SetState(
-        TreeType::vertex_descriptor vertex,
+        DependentStateTreeVertexDescriptor vertex,
         AssetDefs::State newState,
         bool sendNotifications);
   public:
@@ -101,4 +51,4 @@ class StateUpdater
         std::function<bool(AssetDefs::State)> updateStatePredicate);
 };
 
-#endif // ASSETTREE_H
+#endif
