+++ earth_enterprise/src/fusion/autoingest/sysman/memory-tests/SConscript
@@ -37,6 +37,7 @@ env.test('build-raster-project_resourcetest',
             'fusion/autoingest/sysman/AssetVersionD.o',
             'fusion/autoingest/sysman/AssetOperation.o',
             'fusion/autoingest/sysman/StateUpdater.o',
+            'fusion/autoingest/sysman/DependentStateTree.o',
             'fusion/autoingest/sysman/khSystemManager.o',
             'fusion/autoingest/sysman/khAssetManager.o',
             'fusion/autoingest/sysman/khAssetManagerDispatch.o',
