+++ earth_enterprise/src/fusion/autoingest/sysman/parse-raster-project-xml-no-content_unittest/SConscript
@@ -41,6 +41,7 @@ env.test('parse-raster-project-xml-no-content_unittest',
               'fusion/autoingest/sysman/AssetVersionD.o',
               'fusion/autoingest/sysman/AssetOperation.o',
               'fusion/autoingest/sysman/StateUpdater.o',
+              'fusion/autoingest/sysman/DependentStateTree.o',
               'fusion/autoingest/sysman/khSystemManager.o',
               'fusion/autoingest/sysman/khAssetManager.o',
               'fusion/autoingest/sysman/khAssetManagerDispatch.o',
